\name{MySecondRPackage-package}
\alias{MySecondRPackage-package}
\alias{MySecondRPackage}
\docType{package}
\title{
\packageTitle{MySecondRPackage}
}
\description{
\packageDescription{MySecondRPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MySecondRPackage}
\packageIndices{MySecondRPackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{MySecondRPackage}

Maintainer: \packageMaintainer{MySecondRPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
